Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> I am still not sure why the bootloader must load the microcode, but then the kernel is the one to apply it. Why can not the kernel do both, or the bootloader do both?

The kernel would need to wait until after it has mounted the root filesystem and the packages (problematic if we hit a crashing problem before that because old microcode versions have a bug). The bootloader should be kept as simple as possible. So this seems a good compromise to me.

Patch-set: 5
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>