Update patch set 5
Patch Set 5:

You are bound and determined to undermine my work here. Why not build on top of my patch set instead of starting over? Apparently I've stepped on your toes here with regards to your menu layout work.

How about we do it your way and if there's anything left of my patch when you are done we can apply it on top.

You've totally broken my patch set with this change and I'm guessing that was your intentional but all I care about is that the live updating menu feature works at the the end of the day.

Patch-set: 5
CC: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>