Update patch set 5
Patch Set 5:

> Well, X512 had his patch up much before yours, so if anything, I'd say you are kind of stepping on his toes...

3 hours after for this specific patch. Which I assume is accidental, you just both happened to attack the same bug at the same time, and none of you commented on the ticket announcing your intent before doing that.

Anyway, we now have two implementations to chose from. This one is making less changes and possibly safer for BeOS (and even existing Haiku) apps compatibility. We know apps like ProcessController do a lot of unexpected things with menus, and suddenly making the menu use layout kit could be a problem there. It is something we should do in the long term, but it needs careful consideration and testing. I would be worried for my own work in ACE, where I have for example a BSlider inside a menu. But I use B_ITEMS_IN_MATRIX there so at least that should be safe.

So, just communicate more, use the mailing lists and bugtracker to avoid working on the same bug independently :)

Patch-set: 5
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>