Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > This change is needed to fix the regression  in selecting list view items the BeOS way and is ready to vo. Despite humdinger's complaints it is neccessay for list views to update the selection on mouse up instead of mouse down.
> 
> I would go with something like:
> - On mouse down, select the item under the mouse immediately
> - If the user initiates a drag, do the scrolling thing, and update the selection again on mouse up
> 
> I think we get the new behavior as desired, but also preserve the ability to select items on mouse down, then?

I'm going to use the term activated to mean that a list view item changes the display of something and the word selected to mean that the background of the list view item changes. The screen saver list is a good example of this because items can be selected, activated, or invoked.

The way this feature works on BeOS is well thought out (unlike some other parts of BeOS.) It's potentially important that only one list item be selected per mouse down/mouse up. When you push your mouse down on a list view the item is selected but not activated allowing you to move the mouse to another item, select that item, and finally mouse up to activate the item. If you activate on mouse down instead then you break this feature. You could activate on mouse down, then move your mouse to select another item and activate that item too, in fact that's how it worked at one time but that might not work in all cases that a list view might be used in.

Patch-set: 5
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>