Update patch set 5
Patch Set 5:

> > I think we get the new behavior as desired, but also preserve the ability to select items on mouse down, then?
> 
> I am in agreement with PulkoMandy here; basically every other OS I can think of has selection on mouse down, also.

Yes, the behavior of "cancelling" something by holding the mouse button and moving outside the control is nice for actions that are not easily reversible (and even then, it's something that's used very rarely). But on selecting something in a list? I don't see the point. If you selected the wrong item you can always select another one very easily.

Patch-set: 5