Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> > Patch Set 5:
> > 
> > > This change is needed to fix the regression  in selecting list view items the BeOS way and is ready to vo. Despite humdinger's complaints it is neccessay for list views to update the selection on mouse up instead of mouse down.
> > 
> > I would go with something like:
> > - On mouse down, select the item under the mouse immediately
> > - If the user initiates a drag, do the scrolling thing, and update the selection again on mouse up
> > 
> > I think we get the new behavior as desired, but also preserve the ability to select items on mouse down, then?
> 
> I am in agreement with PulkoMandy here; basically every other OS I can think of has selection on mouse down, also.

Not BeOS! What about #15009 ? It's a nice feature and you've regressed it. All I hear is I don't like this change and it feels weird but this is how BeOS worked. I'm not into blindly copying BeOS but this is one area that BeOS got right.

Another big reason to "activate" on mouse up is for drag and drop support, it's the same problem with the same solution as Tracker. We don't want to activate on mouse down because you may click and drag on a background windows list view item to drag it to another list view. If we activate on mouse down the window will come forward potentially getting in the way of your drop target. https://insightfactory.tumblr.com/post/189635203350/beos-r5-tracker-window-drag-and-drop-activation

Patch-set: 5
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>