Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > In Qt5 list selection is updated and immidiately applied during mouse tracking.
> 
> Personally I prefer this behavior, because it allows quickly explore contents that depends on selection.

You mean like this: https://insightfactory.tumblr.com/post/149144871235/beos-had-a-feature-where-you-could-click-on-a-list

Yeah that's one way to do it but my fear is that if activating a list view triggers some long running task like if you have a photo manger app with a list view of albums, it could be an expensive operation to load up the photos in the album. If we activate on scroll that would not work very well. We should make activating on mouse up the default for a list view and provide an option to auto-activate on scroll for the programmer. Then we could turn on the setting on any list views where it is appropriate and get the best of both worlds.

Patch-set: 5
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>