Appearance Prefs: Don't activate window if drag & drop
Set B_WILL_ACCEPT_FIRST_CLICK flag in the APRWindow constructor to stop
the window from automatically activating on mouse down.

In APRWindow::DispatchMessage() Activate() the window when dispatching a
B_MOUSE_DOWN message unless the window is already active or the destination
handler is either ColorWhichListView or ColorPreview.

Activate the window on MouseUp() in ColorWhichListView and ColorPreview.

MouseUp() won't be called if you drag outside the window. This allows you
to drag from the inactive Appearance window to an overlapping active
window without covering the active window with the APRWindow similar to
the same feature in Tracker.

Change-Id: I4f6901e449954576366161264c2a69b1225ea739