Appearance Prefs: Don't activate window if drag & drop
Set B_WILL_ACCEPT_FIRST_CLICK flag in the APRWindow constructor to stop
the window from automatically activating on mouse down.

In APRWindow::DispatchMessage() activate the window unless the window is
already active, the message command is B_MOUSE_DOWN, and the destination
handler is not ColorWhichListView or ColorPreview.

Activate the window on MouseUp in ColorWhichListView and ColorPreview.

MouseUp won't be called if you drag outside the window. This allows you
to drag from the inactive Appearance window to an overlapping active
window without covering the active window with the APRWindow similar to
the same feature in Tracker.