Update patch set 3
Patch Set 3:

> Most list views do not implement drag and drop but this one does. On Haiku things work a bit differently as we have a B_ACTIVATE_ON_FIRST_CLICK flag that must be set to override the default MouseDown behavior in an app.

As I mentionned, we have a setting for this in Mouse preferences and if you want this behavior, you should just use that. It is not something to implement specifically in each app.

Just set your mouse settings to "Click to focus" instead of "Click to focus and raise". And you probably also want to enable "Accept first click" (which makes B_ACCEPT_FIRST_CLICK the default, making such things possible). If you want this, enable it in mouse preferences and keep behavior of all apps and controls consistent. Do not add special cases to each application.

I would even go as far as suggesting we remove the "accept first click" setting completely, but I would like input from other users on that. Do people rely on the behavior of first click only activating the window and not clicking anything in it?

Patch-set: 3