Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> > Patch Set 3:
> 
> > I don't want to change my mouse mode, I've used those settings before and I don't like how they work. I want this feature to work in the normal mode. I don't want a general case I want Appearance to work this way in this specific case, well actually I'd like this to work in any place where drag and drop is used but I could only imagine the howls of I got my wish. I consider the current implementation broken and this is the fix.
> 
> If you personally want it, nobody prevents you apply patch locally and be happy with that, adjusting if necessary when merge conflict appear. Current patch is a hack and not suited to be part of base Haiku. If feature will be implemented in base Haiku, it should be implemented in generic way, adjusting Interface Kit and introducing new API if necessary.

This is broken and needs to be fixed. I already have too many custom changes applied. You call it a hack but this is the API that exists, it's not fair to say that I need to come up with an entirely new API to fix this bug or that I should run custom code. I don't know why I'm getting such push back when I am fixing a bug and this makes the app work much better for every one. It's not a huge deal but I feel that I am being treated unfairly here.

Patch-set: 3
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>