Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This is a hack that I came up with this afternoon so I'm not annoyed about the push back I'm getting on this ticket, but I still like it.

Well, the idea is overall welcome (I would push for focus follows mouse because I think the other mouse modes make little sense precisely because I think it's not possible to get these things working right in the other modes).

However, if we look at the big picture, doing it this way would eventually lead to adding B_ACCEPT_FIRST_CLICK to all windows which can initiate a drag. I see two problems with that:
1) That is pretty muchevery window, because you can drag text from a textfield, colors from a color swatch, and many other things
2) It creates an inconsistency in behavior in unrelated places: any click on the window will now be accepted. This can be dangerous if you expect that your first click will not do anything, but now, on some windows, it does.

And yes, there may be no good way to fix this, unless we activate windows only on mouse up (so we can detect that a mouse operation is in fact a drag, and should not result in activation). I'll let people who use the other mouse settings comment on wether this is appropriate, but in any case, I think the consistency of behavior between all windows is important.

Patch-set: 3