Update patch set 1
Patch Set 1:

(1 comment)

I proposed this change a while back and it was rejected, let's see what happens when it's not me proposing the change.

What is the friend class change in Menu.h needed for?

Assuming that this change is deemed acceptable by the powers that be, the documentation in docs/user/interface/SeparatorIem.dox should also be updated. There is a note "BSeparatorItems are only meant to be used with menus whose items are arranged in a B_ITEMS_IN_COLUMN layout." that should be altered to reflect the updated ability of BSeparator class https://www.haiku-os.org/docs/api/classBSeparatorItem.html

My suggested for the change in documentation is as follows:

BSeparatorItems are intended to be used with menus whose items are arranged in \c B_ITEMS_IN_COLUMN or \c B_ITEMS_IN_ROW layout.

Patch-set: 1
CC: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>