Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> It may be good to move BMenu::Layout() from protected to public, beacuse BMenuItem may need to know in which layout it is displayed.

If we were to make that kind of change we should name it BMenu::MenuLayout() (and BMenu::SetMenuLayout()) because the Layout() name conflicts with BView::Layout() introduced with the Layout APIs. This change should be added to the list of changes we will make as soon as we have decided to break with BeOS binary compatibility.

Use BPrivate::MenuPrivate in the BSeparatorItem classs and take the friend class out of Menu.h and then patch set will be ready to go.

Patch-set: 2