Update patch set 2
Patch Set 2:

> I proposed this change a while back and it was rejected, let's see what happens when it's not me proposing the change.

Can we stop the conspiracy theories, please?

> Assuming that this change is deemed acceptable by the powers that be

You have as much decision power as anyone else here. If you can find one person who agrees with you and submit your changes, and no one gives a -2 for huge problems (or "over my dead body" as we say sometime), you can get things merged.

The review process is not for personal attacks or so. It's for discussing possibly better ideas or implementations, and making sure that the overall consistency and behavior is preserved.

I tend to be quite conservative on little details. But you can handle me as the old (or not that old) man always complaining that things were better in earlier times: ignore me and make changes anyways. I rarely give a -2 to changes for such reasons, unless I see a very large flaw.

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>