Update patch set 2
Patch Set 2:

On the code review: I have nothing to add to John concerns and suggestions.

On the functionality: I see no problem in having the separators work in horizontal menus, it makes sense to have this working.

However, there is the wider questions:

1) Do we need toolbars at all?

This was debated when John first suuggested such changes, I think. I probably said we didn't back then. The arguments are things like:
- Toolbars with icons only allow quick access to functionalities and in a quite compact way (assuming smaller icons). But it is hard to convey the meaning of a button with just an icon.
- Toolbars use up vertical space, which is precious on current 16:9 screens (fortunately 21:9 is only for phones, for now at least). This is worse if the icons are large (and the current ones are larger than usual)

However, it seems the general agreement is that, while being aware of these limitations, there are still some good uses for toolbars.

As you know, meanwhile Windows went the opposite direction, and reworked both toolbars and menus into the new "ribbon" interface. I still don't think this is good, because it still wastes about as much vertical spaces, and is not as discoverable as menus (still tries to convey too much meaning through icons).

I have no better idea for this, anyway, so, let's say we're ok with toolbars used wisely.

2) Should toolbars be menus or a separate class?

The current answer is that we have a BToolBar. But there are counter-examples, the bookmark bar in Web+ is implemented as a BMenu (and can have submenus). Vision has a regular menu of which one item has no label and only the application icon, which I think works well (it is obvious that this is an "Application"/"Main" menu). On the other hand, apps using BToolBar (for example, Slayer), indeed look like they are wasting a lot of screen space for not much.

So, should we deprecate BToolBar and go with icon in menus instead? Is there a need for both? Should we allow both and let application writers decide? I don't know.

Regardless, allowing the separator view to work in horizontal menus still sounds like a good thing. Wether we encourage this in the HIG, and actually use it in our apps, is another debate.

Patch-set: 2