Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Building haiku.hpkg with this patch works, haiku_x86.hpkg fails (32bit, obviously):
> 
> src/kits/interface/ListView.cpp: In member function 'virtual void BListView::MouseDown(BPoint)':
> src/kits/interface/ListView.cpp:586:13: error: suggest parentheses around comparison in operand of '&' [-Werror=parentheses]
>     = (index >= 0 & item != NULL ? item->IsSelected() : false);
>        ~~~~~~^~~~
> src/kits/interface/ListView.cpp: In member function 'bool BListView::_Select(int32, bool)':
> src/kits/interface/ListView.cpp:1544:19: error: suggest parentheses around '&&' within '||' [-Werror=parentheses]
>   if (item != NULL && !item->IsEnabled() || item->IsSelected()) {

The first error was because I used a bitwise & instead of a logical && and is fixed in patch set 7... I hope. According to my understanding of C++ operator precedence rules this should now work without any additional parens.

The second warning turned to error was a legitimate lack of parentheses fixed in patch set 8.

Thanks for trying to test my patch.

Patch-set: 6