Tracker: Fix popup MenuEnded() cleanup.
Set fContextMenu to the right variable in ShowContextMenu() and call
fContextMenu->Go() at the end. The only time that we exit this method
without calling fContextMenu->Go() is if we are Dragging() a file or
if the context menu is already open. With fContextMenu set we can check
it in MenusEnded() and if set exit skipping all the code that is intended
for the menus in the menu bar like the File menu.

Create fInitStatus and check it instead of checking each variable in
SetupMoveCopyMenus() and also check fInitStatus in ShowContextMenu().