Update patch set 2
Patch Set 2:

> > Patch Set 1:
> > 
> > This doesn't work at all. If you're not testing on top of my Tracker live updating menu branch you won't be able to test this change properly.
> 
> This is working fine with test program. You are doing something wrong in your patch, probably manually resizing BMenuWindow, that is dirty hack.

First of all I took out the manually resizing BMenuWindow (that you call a dirty hack) already, second of all, I had tested what this patch does on my own and so I know that _UpdateWindowViewSize() doesn't get called at the right time for this feature to work, third of all I tested your patch with Tracker live-shifting menus and opening a submenu prevents the menu resizing from working, and lastly if you're test program is not Tracker with live-shifting menus I do not want to hear about it. There are several intricacies that are handled by the menu window layout that your code does not address.

Patch-set: 2