Update patch set 1
Patch Set 1:

This patch seems to have fixed the problem, you got in now (you that you had to test on top of live-shifting tracker branch to see the problem.) What's the secret? I've abandoned my menu window change now because this patch and 2324 succeeds it.

TLDR; there is a minor bug: https://i.imgur.com/uJz5k04.png

In my MoveSubmenusOver() method I reproduced the menu window layout used by BMenu exactly so that when you pressed shift you wouldn't notice the difference. While doing this I reproduced a BMenu bug. The bug is a bit tough to trigger but you can reproduce the issue in a context click file menu in Tracker (my live shifting changes are not required for this.) All you have to do is open enough submenus to go all the way to the right side of the screen, when it hits the right side it starts opening menus to the left, that part works, but when you get all the way to the left side of the screen submenus no longer layout correctly. It should make the submenus popup back towards the right side of the screen again. It's a lot easier to trigger if you start the context menus on the right side of the screen and make your screen resolution smaller. I would categorize this bug as extremely minor but if you can get around to fixing it I'd appreciate the fix.

Patch-set: 1
CC: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>