Update patch set 6
Patch Set 6:

> I'm curious, how is the BPopUpMenu case different from the BMenuBar case?

Because, from the API point of view, the popup menu is not related to a particular window. You can create one without having a window, even. And, as X512 mentionned, they are invoked directly by the user.

It's even documented in the Be Book:

The MenusBeginning() hook function is called just before menus *belonging to the window* are about to be shown to the user.

A popup menu does not belong to the window. It just happens to be invoked from there.

I think the best way to do this would be to call MenusBeginning from BMenuBar (which is also used in BMenuField and BOptionPopUp), because BMenuBar is the one which makes a menu belong to a window.

For DeskBar, what I would do is simply modify replicants to call MenusBeginning/MenusEnded themselves when they spawn a menu. They are the one who make the menu "belong" to the window, and I think trying to infer that from inside the popup menu is not a good idea.

Patch-set: 6