Update patch set 23
Patch Set 23:

It really must be possible to word the commit message in less words. I appreciate that you take the time to explain everything, but you should never come to the point asking whether or not the commit message looks better block justified or left aligned.

Either break the commit into smaller ones, or post an additional blog post that delivers this level of detail.

Patch-set: 23
CC: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>