Update patch set 23
Patch Set 23:

> Patch Set 23:
> 
> It really must be possible to word the commit message in less words. I appreciate that you take the time to explain everything, but you should never come to the point asking whether or not the commit message looks better block justified or left aligned.
> 
> Either break the commit into smaller ones, or post an additional blog post that delivers this level of detail.

I will endeavour to do better in the future. The commit message length is based on the amount of work that went into this patch set which was considerable. I put some blocks of code into the commit message that got removed because there might be some hidden gems of knowledge in the old code that was lost and a code archaeologist (possibly myself) may be aided by these details in the future. Getting the replicant icons to go on the side was not that hard, fixing the bugs after was.

I would like to make smaller patches but every time I split up changes it adds overhead that makes my life harder. I split out the BWindow change into its own commit because it was logically separate from Deskbar and that caused a great deal of frustration for me.

With this patch applied so many small Deskbar bugs have been fixed that it will be easier to make changes in the future without breaking things which should produce smaller commit messages. Everything was slightly off and it was hard to notice but to get this change past the sensors I had to fix a ton of bugs. Mode switching was subtly broken. The status tray was the wrong size and location. There was unseen drawing code, the Team menu was badly layed out, there were bugs with the scroller arrows.

Consequently this touched so many lines of code that future Deskbar developers are bound to hit this commit in git blame and I want them to be able to see all these details so they know what changed and why.

Patch-set: 23