Update patch set 1
Patch Set 1:

> It is not too destructive, for non-saved documents alert will be displayed before closing.

It is destructive in the sense that it can unintentionally close a window with a lot of context (for example, a web browser window with many tabs). No document is lost, but the context is. Likewise for Terminal window with bash history and multiple tabs, image viewer where you had loaded a specific image and zoomed on a specific area, etc.

> Middle click is often used in web browsers for closing tabs. I think that there a lot of people are familiar with this shortcut. Also I don't think that it is easy to perform middle button click accidentally, at least with properly designed mouse.

What happens often for me is that I try to paste something using middle button, but I accidentally click on the wrong place and end up closing, for example, a terminal tab, instead of pasting in terminal.

In Deskbar, it was recently suggested that middle click on Tracker (app icon) could create a new window (as is the case on Windows). If we are going to make DeskBar a more dock-like thing, that would be a good idea, but then it raises the risk that such a mis-click could happen.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>