Update patch set 1
Patch Set 1:

(2 comments)

For the crash you have linked, a debug report would be more useful than a screenshot. You can get one by typing save-report at the prompt, it will be saved on the desktop.

It looks like you hit an ASSERT_PARALLEL_LOCKED assert in the DrawingEngine, but I don't know why this would be. Possibly a font ended up trying to fallback to itself or something like that?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>