bfs: fix misaligned access
The BFS on-disk data is not aligned. Reading it to memory and trying to
access fields directly does not work on sparc. memcpy the data to an
aligned variable before handling it with its native size.

There is a performance regression in doing byte access on all
architectures, and I'm not sure how well gcc will optimize this. It
should be possible to use #ifdef to get most of the code to still use
the previous access methods (done in ValueAt here, but not in the other
accessors). It probably needs a more wider #ifdef than just sparc, as
apparently ARM (old versions, at least) would have the same restriction.

Fixes #9255

Change-Id: I3bf62590dee059ad32b1845bdc4eace165b73203