Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Can you explain what you mean by "The only difference is when the size can be divided without rest."? I'm having trouble understanding this bit. I don't know what divided without rest means. Division is happening here width / count but what's rest?
> 
> Sorry, I meant remainder ("Rest" is the german word for it).

I feel a little bit dumb for not getting this, but I looked up the term and rest is never a synonym for remainder in English, but it is uncommonly used to refer to the quotient, i.e. the result of division not including the remainder.

I’ll make whatever change is necessary to get this change in but from my understanding the code I wrote is correct.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>