Update patch set 2
Patch Set 2:

It does not help with the core of the problem.

The initial problem is that for sume USB transfers, the completion callback is not called. The transfers fails or stalls or something at the USB level, but the callback is not called.

In usb_raw, that callback is used to unblock the userland thread that's waiting for the transfer.

So, now, we unlock that when the device is removed (for example, if you unplug it). But we shouldn't get in that situation in the first place, if a transfer fails, the syscall should return with an error immediately, not wait until the device is removed.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>