Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (1 comment)

I disagree with this assessment.  This situation where this problem arises is very much a "system error" scenario.  It's not normal or expected that the standard cache directory is unable to be written to and warning the user in this way seems entirely appropriate.  It would be unreasonable an onerous to create the application to be function with as well as without functional storage on hand.

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>