Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> 8. Default button label is not horizontally centered, see "Update now", "Keep enabled".

This is true on HaikuControlLook as well but you’re right that regular buttons start 1px lower than the effective area of default buttons and this didn’t happen on BeOS R5. So this is a bug, just not a bug in BeControlLook, it’s in BButton. *sigh*

Patch-set: 4
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>