Update patch set 2
Patch Set 2: Code-Review+1

Same with this one, would prefer a newline before return. You put a newline on one method but not the other. This is my personal preference and is not part of the style guidelines so you may feel free to ignore this preference of mine if you wish.

Patch-set: 2
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1