HaikuDepot: Check User Auth on Start
The user might have changed their authentication details
on the server and the client won't detect this until
they go to do something.  Instead, if possible, check
this as the client starts.  Also check that the user has
agreed to the current user usage conditions.

As a side-effect this generalizes the logic for process
coordination in the main window and also fixes some bugs
in the main window's progress display as the application
starts.

Relates to #15209