Update patch set 2
Patch Set 2:

Hi!

I think this is not the right thing to do. If the device exists, the Input preferences should show it.

The question is, why does this device exist, when there is no tablet connected to the computer?

You should look at the function that enumerates the devices. Where do they come from in input_server? Why is the wacom input_server add-on adding a device when there is no tablet connected? Finding this will allow to fix the problem at the source.

Patch-set: 2
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>