Update patch set 1
Patch Set 1: Code-Review-1

This is not in POSIX but is a GNU extension that exists only in the glibc. So the declaration should go in headers/compatibility/gnu and not headers/posix. And maybe we don't want to implement every GNU extension, too. What's the use case for this one?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1