Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> This is not in POSIX but is a GNU extension that exists only in the glibc. So the declaration should go in headers/compatibility/gnu and not headers/posix. And maybe we don't want to implement every GNU extension, too. What's the use case for this one?

FreeBSD has qsort_r also.

This is useful when your comparison function needs extra state for some reason, for which there are a pretty wide variety of cases. Actually libsolv needs it, but it has a shim in case it isn't provided which is why we have not had an issue with that.

Patch-set: 1
CC: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>