i2c_hid: driver for i2c-hid devices
This is getting more common in tablets and laptops. It replaces PS/2 for
the internal keyboard and pointing devices. This is simpler and cheaper
than using up USB ports, and also simpler than the old and quirky PS/2
protocol.

The HID spec is the same no matter what transport is used (it is also
applicable for Bluetooth). Ideally we could create a separate HID bus
manager that would handle all these devices in a generic way, but that
is a lot of work nad extra complications for uncertain gains.

For now, just move the common files to a shared directory where both
drivers can use them. As a result the files are compiled twice, which is
what we want, because currently they hardcode some device paths that
need to be different for each driver.

Some files are copied between the usb and i2c drivers with slight
changes. If this approach seems acceptable, I can try to make the needed
changes to have these files shared again:
- HIDReport: minor API change in how the driver is interfaced
- Keyboard protocol: access to low level USB info for KDL, not
  applicable to the i2c version

Change-Id: I0327f6864dd0a4372b708f7b7ecf299aa86a6ea9