Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Which applications for instance?
> 
> Why does it matter? (Actually I don't know if app_server even does this correctly all the time.)

It matters, why include a typeface with a country code, causing confusion?

Why prioritise Japanese users, instead of doing it universally?

If an OS prioritises its behaviour to fix some application behaviour, I'd like to know which application does so, so that we can weigh it if it's worth it or not.

Patch-set: 1
Reviewer: Gerrit User 1000169 <1000169@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>