Update patch set 1
Patch Set 1:

> It matters, why include a typeface with a country code, causing confusion?

Why does it cause confusion? You should never see this typeface's name (except in the installed font list I guess) unless you are actually using it.

> Why prioritise Japanese users, instead of doing it universally?

Because I don't think we handle the necessary switching at Be API and app_server level, let alone in all applications, and so we decided to prioritize the users we actually have as the default.

We asked at the time for the CJK community, or anyone with the requisite knowledge, to help us out with this, but nobody did (and still has not). So, if the necessary changes are made, and legacy apps are still supported properly (I doubt this criterion can be fulfilled, though), then we can switch to the non-"JP" variant. But we absolutely cannot in the current state of things.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>